Rename mark status button file, Fix imports
